Move the custom error type to its own module
